+++ client/branded/src/search-ui/components/OwnerSearchResult.tsx
@@ -4,7 +4,11 @@ import classNames from 'classnames'
 
 import { TeamAvatar } from '@sourcegraph/shared/src/components/TeamAvatar'
 import { UserAvatar } from '@sourcegraph/shared/src/components/UserAvatar'
+import { BuildSearchQueryURLParameters, QueryState, SearchContextProps } from '@sourcegraph/shared/src/search'
+import { FilterKind, findFilter } from '@sourcegraph/shared/src/search/query/query'
+import { appendFilter, omitFilter } from '@sourcegraph/shared/src/search/query/transformer'
 import { getOwnerMatchUrl, OwnerMatch } from '@sourcegraph/shared/src/search/stream'
+import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { Link } from '@sourcegraph/wildcard'
 
 import { ResultContainer } from './ResultContainer'
@@ -12,20 +16,28 @@ import { ResultContainer } from './ResultContainer'
 import styles from './OwnerSearchResult.module.scss'
 import resultStyles from './SearchResult.module.scss'
 
-export interface PersonSearchResultProps {
+export interface OwnerSearchResultProps extends Pick<SearchContextProps, 'selectedSearchContextSpec'>, TelemetryProps {
     result: OwnerMatch
     onSelect: () => void
     containerClassName?: string
     as?: React.ElementType
     index: number
+
+    // If not provided, the result will not contain a link to the owner's files.
+    queryState?: QueryState
+    buildSearchURLQueryFromQueryState?: (queryParameters: BuildSearchQueryURLParameters) => string
 }
 
-export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps> = ({
+export const OwnerSearchResult: React.FunctionComponent<OwnerSearchResultProps> = ({
     result,
     onSelect,
     containerClassName,
     as,
     index,
+    queryState,
+    buildSearchURLQueryFromQueryState,
+    selectedSearchContextSpec,
+    telemetryService,
 }) => {
     const displayName = useMemo(() => {
         let displayName = ''
@@ -40,12 +52,49 @@ export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps>
 
     const url = useMemo(() => {
         const url = getOwnerMatchUrl(result)
-        if (result.type === 'person' && !result.user) {
+        const validUrlPrefixes = ['/teams/', '/users/', 'mailto:']
+        if (!validUrlPrefixes.some(prefix => url.startsWith(prefix))) {
             // This is not a real URL, remove it.
             return ''
         }
         return url
     }, [result])
+
+    const fileSearchLink = useMemo(() => {
+        if (!queryState || !buildSearchURLQueryFromQueryState) {
+            return ''
+        }
+
+        const handle = result.handle || result.email
+        if (!handle) {
+            return ''
+        }
+
+        let query = queryState.query
+        const selectFilter = findFilter(queryState.query, 'select', FilterKind.Global)
+        if (selectFilter && selectFilter.value?.value === 'file.owners') {
+            query = omitFilter(query, selectFilter)
+        }
+        query = appendFilter(query, 'file', `has.owner(${handle})`)
+
+        const searchParams = buildSearchURLQueryFromQueryState({
+            query,
+            searchContextSpec: selectedSearchContextSpec,
+        })
+        return `/search?${searchParams}`
+    }, [buildSearchURLQueryFromQueryState, queryState, result.email, result.handle, selectedSearchContextSpec])
+
+    const logSearchOwnerClicked = (): void => {
+        if (url.startsWith('mailto:')) {
+            telemetryService.log('searchResults:ownershipMailto:clicked')
+        }
+        if (url.startsWith('/users/')) {
+            telemetryService.log('searchResults:ownershipUsers:clicked')
+        }
+        if (url.startsWith('/teams/')) {
+            telemetryService.log('searchResults:ownershipTeams:clicked')
+        }
+    }
 
     const title = (
         <div className="d-flex align-items-center">
@@ -68,7 +117,7 @@ export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps>
             )}
 
             {url ? (
-                <Link to={url} className="text-muted">
+                <Link to={url} className="text-muted" onClick={logSearchOwnerClicked}>
                     {displayName}
                 </Link>
             ) : (
@@ -91,9 +140,14 @@ export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps>
                 className={classNames(resultStyles.searchResultMatch, 'p-2 flex-column')}
                 data-testid="owner-search-result"
             >
-                <div className={resultStyles.matchType}>
-                    <small>Owner match</small>
-                </div>
+                <small className={resultStyles.matchType}>
+                    <span>Owner match</span>
+                    {fileSearchLink && (
+                        <Link to={fileSearchLink} className={styles.filesLink}>
+                            Show files
+                        </Link>
+                    )}
+                </small>
                 {result.type === 'person' && !result.user && (
                     <>
                         <div className={resultStyles.dividerVertical} />