+++ enterprise/cmd/frontend/internal/own/resolvers/resolvers_test.go
@@ -11,6 +11,8 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/graph-gophers/graphql-go/relay"
 
+	"github.com/sourcegraph/log/logtest"
+
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/own/resolvers"
@@ -95,6 +97,7 @@ func (g fakeGitserver) Stat(ctx context.Context, checker authz.SubRepoPermission
 // query, where the owner is unresolved. In that case if we have a handle, we only return
 // it as `displayName`.
 func TestBlobOwnershipPanelQueryPersonUnresolved(t *testing.T) {
+	logger := logtest.Scoped(t)
 	fs := fakedb.New()
 	db := database.NewMockDB()
 	fs.Wire(db)
@@ -119,7 +122,7 @@ func TestBlobOwnershipPanelQueryPersonUnresolved(t *testing.T) {
 		return "42", nil
 	}
 	git := fakeGitserver{}
-	schema, err := graphqlbackend.NewSchema(db, git, nil, graphqlbackend.OptionalResolver{OwnResolver: resolvers.New(db, git, own)})
+	schema, err := graphqlbackend.NewSchema(db, git, nil, graphqlbackend.OptionalResolver{OwnResolver: resolvers.New(db, git, own, logger)})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -355,6 +358,7 @@ func (r paginationResponse) ownerNames() []string {
 // *  all results are eventually returned, in the expected order;
 // *  each request returns correct pageInfo and totalCount;
 func TestOwnershipPagination(t *testing.T) {
+	logger := logtest.Scoped(t)
 	fs := fakedb.New()
 	db := database.NewMockDB()
 	fs.Wire(db)
@@ -383,7 +387,7 @@ func TestOwnershipPagination(t *testing.T) {
 		return "42", nil
 	}
 	git := fakeGitserver{}
-	schema, err := graphqlbackend.NewSchema(db, git, nil, graphqlbackend.OptionalResolver{OwnResolver: resolvers.New(db, git, own)})
+	schema, err := graphqlbackend.NewSchema(db, git, nil, graphqlbackend.OptionalResolver{OwnResolver: resolvers.New(db, git, own, logger)})
 	if err != nil {
 		t.Fatal(err)
 	}